requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymongo
email-validator
pydantic-settings
pytest
httpx
pytest-asyncio
mongomock